[MOSIP-43633] Code changes for removing the cyclic dependency from the biometric update bug fix code.
