+++ registration-processor/registration-processor-info-storage-service/src/main/java/io/mosip/registration/processor/packet/storage/utils/Utilities.java
@@ -2,19 +2,13 @@ package io.mosip.registration.processor.packet.storage.utils;
 
 import java.io.IOException;
 import java.time.*;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
-import io.mosip.kernel.biometrics.commons.CbeffValidator;
-import io.mosip.kernel.biometrics.entities.BIR;
-import io.mosip.kernel.biometrics.entities.BiometricRecord;
-import io.mosip.kernel.core.util.CryptoUtil;
-import io.mosip.registration.processor.core.constant.*;
 import io.mosip.registration.processor.core.exception.*;
-import io.mosip.registration.processor.core.idrepo.dto.*;
 import io.mosip.registration.processor.core.packet.dto.AdditionalInfoRequestDto;
-import io.mosip.kernel.biometrics.constant.BiometricType;
-import io.mosip.registration.processor.status.entity.SyncRegistrationEntity;
-import io.mosip.registration.processor.status.repositary.SyncRegistrationRepository;
 import io.mosip.registration.processor.status.service.AdditionalInfoRequestService;
 import org.apache.commons.lang.StringUtils;
 import org.json.simple.JSONArray;
@@ -29,13 +23,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.mosip.kernel.core.exception.ExceptionUtils;
 import io.mosip.kernel.core.logger.spi.Logger;
-import io.mosip.kernel.core.util.exception.JsonProcessingException;
 import io.mosip.registration.processor.abis.queue.dto.AbisQueueDetails;
 import io.mosip.registration.processor.core.code.ApiName;
 import io.mosip.registration.processor.core.common.rest.dto.ErrorDTO;
 import io.mosip.registration.processor.core.constant.LoggerFileConstant;
-import io.mosip.registration.processor.core.constant.MappingJsonConstants;
-import io.mosip.registration.processor.core.constant.ProviderStageName;
 import io.mosip.registration.processor.core.exception.util.PlatformErrorMessages;
 import io.mosip.registration.processor.core.idrepo.dto.IdResponseDTO1;
 import io.mosip.registration.processor.core.idrepo.dto.ResponseDTO;
@@ -54,7 +45,6 @@ import io.mosip.registration.processor.status.dao.RegistrationStatusDao;
 import io.mosip.registration.processor.status.dto.InternalRegistrationStatusDto;
 import io.mosip.registration.processor.status.entity.RegistrationStatusEntity;
 import lombok.Data;
-import static io.mosip.kernel.core.util.DateUtils.*;
 
 /**
  * The Class Utilities.
@@ -73,8 +63,6 @@ public class Utilities {
 	private static final String SOURCE = "source";
 	private static final String PROCESS = "process";
 	private static final String PROVIDER = "provider";
-	public static final String EXCEPTION = "EXCEPTION";
-	public static final String TRUE = "TRUE";
 
 	private static Map<String, String> readerConfiguration;
 	private static Map<String, String> writerConfiguration;
@@ -114,12 +102,6 @@ public class Utilities {
 	@Autowired
 	private MosipQueueConnectionFactory<MosipQueue> mosipConnectionFactory;
 
-	@Autowired
-	SyncRegistrationRepository<SyncRegistrationEntity, String> syncRegistrationRepository;
-
-	@Autowired
-	Utility utility;
-
 	@Value("${provider.packetreader.mosip}")
 	private String provider;
 
@@ -138,10 +120,6 @@ public class Utilities {
 	/** The get reg processor applicant type. */
 	@Value("${registration.processor.applicant.type}")
 	private String getRegProcessorApplicantType;
-
-	/** The dob format. */
-	@Value("${registration.processor.applicant.dob.format}")
-	private String dobFormat;
 
 	/** The elapse time. */
 	@Value("${registration.processor.reprocess.elapse.time}")
@@ -164,41 +142,6 @@ public class Utilities {
 
 	@Value("#{'${registration.processor.main-processes}'.split(',')}")
 	private List<String> mainProcesses;
-
-	/**
-	 * Configuration property that defines the age limit used to determine
-	 * whether an applicant is considered an infant or non-infant.
-	 */
-	@Value("${mosip.kernel.applicant.type.age.limit:5}")
-	private String ageLimit;
-
-	/**
-	 * Buffer (in months) applied to the applicant's age limit to handle edge cases
-	 * where applicants are close to the threshold. Default is {@code 0}.
-	 *
-	 * <p>Increasing this value provides a safety margin for age calculations near
-	 * the eligibility boundary.</p>
-	 */
-	@Value("${registration.processor.applicant.type.age.limit.buffer:0}")
-	private Integer ageLimitBuffer;
-
-	/**
-	 * Estimated time (in hours) the system takes to process a packet.
-	 * <p>
-	 * This value helps calculate the approximate packet creation time.
-	 * It is used when no other data source is available to determine the
-	 * packet creation time, apart from the identity update time.
-	 * The system subtracts this duration from the identity update time
-	 * to estimate when the packet was likely created, since the identity
-	 * update time includes the packet processing duration as well.
-	 * The default value is {@code 0}.
-	 * </p>
-	 */
-	@Value("${registration.processor.expected-packet-processing-duration:0}")
-	private Integer expectedPacketProcessingDurationHours;
-
-	@Autowired
-	private PriorityBasedPacketManagerService packetManagerService;
 
 	/** The registration status dao. */
 	@Autowired
@@ -595,7 +538,7 @@ public class Utilities {
 	 * @return the latest transaction id
 	 */
 	public String getLatestTransactionId(String registrationId, String process, int iteration,
-			String workflowInstanceId) {
+										 String workflowInstanceId) {
 		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
 				registrationId, "Utilities::getLatestTransactionId()::entry");
 		RegistrationStatusEntity entity = registrationStatusDao.find(registrationId, process, iteration,
@@ -780,7 +723,7 @@ public class Utilities {
 		return centerId + "_" + machineId;
 	}
 
-public String getInternalProcess(Map<String, String> additionalProcessMap, String externalProcess){
+	public String getInternalProcess(Map<String, String> additionalProcessMap, String externalProcess){
 		if (externalProcess == null) return "";
 		String internalProcess = additionalProcessMap.get(externalProcess);
 		return internalProcess != null ? internalProcess : "";
@@ -795,607 +738,5 @@ public String getInternalProcess(Map<String, String> additionalProcessMap, Strin
 			throw new AdditionalInfoIdNotFoundException();
 
 		return additionalInfoRequestDto.getAdditionalInfoIteration();
-	}
-
-	/**
-	 * Determines whether the applicant was an infant at the time their last packet was processed in the system
-	 *
-	 * @param registrationId
-	 * @param registrationType
-	 * @param stageName
-	 * @return
-	 * @throws ApisResourceAccessException
-	 * @throws IOException
-	 * @throws JsonProcessingException
-	 * @throws PacketManagerException
-	 */
-	public boolean wasInfantWhenLastPacketProcessed(String registrationId, String registrationType, ProviderStageName stageName) throws ApisResourceAccessException, IOException,
-			JsonProcessingException, PacketManagerException {
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
-				registrationId, "utility::wasInfantWhenLastPacketProcessed()::entry");
-
-		String packetUin = utility.getUIn(registrationId, registrationType, stageName);
-		if (packetUin == null || packetUin.trim().isEmpty()) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-					LoggerFileConstant.REGISTRATIONID.toString(),
-					registrationId,
-					"UIN not found in the packet");
-			return false;
-		}
-		// Get created date and time from idrepo using above UIN
-		JSONObject idvidResponse = idRepoService.getIdJsonFromIDRepo(packetUin, getGetRegProcessorDemographicIdentity());
-
-		LocalDate lastPacketProcessedDate = resolveLastPacketProcessedDate(registrationId, packetUin, idvidResponse);
-
-		if (lastPacketProcessedDate == null) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"Unable to compute the creation date of the last processed packet");
-			throw new PacketDateComputationException(PlatformErrorMessages.RPR_BDD_UNABLE_TO_COMPUTE_CREATION_DATE.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_COMPUTE_CREATION_DATE.getMessage());
-		}
-
-		LocalDate dobOfApplicant = getDateOfBirthFromIdrepo(registrationId, idvidResponse);
-		if (dobOfApplicant == null) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"Unable to obtain the date of birth");
-			throw new PacketDateComputationException(PlatformErrorMessages.RPR_BDD_UNABLE_TO_COMPUTE_CREATION_DATE.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_COMPUTE_CREATION_DATE.getMessage());
-		}
-		int age = calculateAgeAtLastPacketProcessing(dobOfApplicant, lastPacketProcessedDate, registrationId);
-
-		regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
-				registrationId, "utility::wasInfantWhenLastPacketProcessed()::exit. age: {}", age);
-
-		return age < getEffectiveAgeLimit();
-	}
-
-
-	/**
-	 * Attempts to resolve the last packet processed date using multiple strategies in order
-	 *
-	 * @param registrationId
-	 * @param packetUin
-	 * @param idvidResponse
-	 * @return
-	 * @throws ApisResourceAccessException
-	 * @throws IOException
-	 */
-	private LocalDate resolveLastPacketProcessedDate(String registrationId, String packetUin, JSONObject idvidResponse) throws ApisResourceAccessException, IOException {
-		// 1. Try direct lookup
-		LocalDate date = getLastProcessedPacketCreatedDate(registrationId, idvidResponse);
-		if (date != null) {
-			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"resolveLastPacketProcessedDate() :: Successfully resolved packet creation date from idvidResponse. date : {}", date);
-			return date;
-		}
-		// Fetching last processed rid from idVidMetadata
-		IdVidMetadataResponse idVidMetadataResponse = getIdVidMetadata(packetUin, null);
-		if (idVidMetadataResponse == null) {
-			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"resolveLastPacketProcessedDate() :: idVidMetadataResponse is null");
-			return null;
-		}
-
-		// 2. Try from packetId (yyyyMMddHHmmss)
-		date = getPacketCreatedDateFromSyncRegistration(idVidMetadataResponse.getRid());
-		if (date != null) {
-			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"resolveLastPacketProcessedDate() :: Successfully resolved packet creation date from packetId. date : {}", date);
-			return date;
-		}
-
-		// 3. Try from RID directly (yyyyMMddHHmmss)
-		date = getPacketCreatedDateFromRid(idVidMetadataResponse.getRid());
-		if (date != null) {
-			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"resolveLastPacketProcessedDate() :: Successfully resolved packet creation date from RID. date : {}", date);
-			return date;
-		}
-
-		// 4. Fallback to IdRepo update date
-		LocalDate approxCreatedDateTime = computePacketCreatedFromIdentityUpdate(idVidMetadataResponse, registrationId);
-		if (approxCreatedDateTime != null) {
-			regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-					"resolveLastPacketProcessedDate() :: Resolved packet creation date from identity update. approxCreatedDateTime : {}", approxCreatedDateTime);
-			return approxCreatedDateTime;
-		}
-
-		regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), registrationId,
-				"resolveLastPacketProcessedDate() :: Unable to resolve packet creation date");
-		return null;
-	}
-
-	 /**
-	 * Retrieves the effective age limit for the applicant by adding the configured buffer (in months)
-	 * to the base age limit. This helps prevent exclusion of applicants who are close to the age threshold
-	 * due to edge cases in age calculation
-	 */
-	public int getEffectiveAgeLimit() {
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), "",
-				"utility::getEffectiveAgeLimit():: ageLimit: {}, ageLimitBuffer: {} ", ageLimit, ageLimitBuffer);
-		return Integer.parseInt(ageLimit) + ageLimitBuffer;
-	}
-
-	/**
-	 * Retrieves the created date of the last packet that was processed for the applicant
-	 *
-	 * @param rid
-	 * @param idvidResponse
-	 * @return
-	 * @throws IOException
-	 */
-	private LocalDate getLastProcessedPacketCreatedDate(String rid, JSONObject idvidResponse) throws IOException {
-
-		String packetCreatedDateTimeIsoFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getLastProcessedPacketCreatedDate()::entry");
-
-		String packetCreatedOnFieldName = getMappedFieldName(MappingJsonConstants.PACKET_CREATED_ON);
-
-		// Check if the response object itself is null
-		if (idvidResponse == null) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"getLastProcessedPacketCreatedDate() :: idvidResponse is null");
-			return null;
-		}
-
-		// Check if the key exists in the response
-		if (!idvidResponse.containsKey(packetCreatedOnFieldName)) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"getLastProcessedPacketCreatedDate() :: fieldName does not exist in idvidResponse : {} ", packetCreatedOnFieldName);
-			return null;
-		}
-
-		// Safely get the value
-		String packetCreatedOn = JsonUtil.getJSONValue(idvidResponse, packetCreatedOnFieldName);
-
-		// Check if the value itself is null
-		if (packetCreatedOn == null) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"getLastProcessedPacketCreatedDate() :: fieldValue does not exist in idvidResponse : {} ", packetCreatedOnFieldName);
-			return null;
-		}
-		LocalDate packetCreatedDate = parseToLocalDate(packetCreatedOn, packetCreatedDateTimeIsoFormat);
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getLastProcessedPacketCreatedDate()::exit");
-		return packetCreatedDate;
-	}
-
-	public String getMappedFieldName(String key) throws IOException {
-		JSONObject regProcessorIdentityJson = getRegistrationProcessorMappingJson(MappingJsonConstants.IDENTITY);
-		return JsonUtil.getJSONValue(
-				JsonUtil.getJSONObject(regProcessorIdentityJson, key),
-				MappingJsonConstants.VALUE);
-
-	}
-
-	/**
-	 * Extract Date of Birth from IDRepo using RID
-	 *
-	 * @param rid
-	 * @param idvidResponse
-	 * @return
-	 * @throws IOException
-	 */
-	public LocalDate getDateOfBirthFromIdrepo(String rid, JSONObject idvidResponse) throws IOException {
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getDateOfBirthFromIdrepo()::entry");
-
-		// Step 2: Fetch DOB dynamically via mapping
-		String dateOfBirthFieldName = getMappedFieldName(MappingJsonConstants.DOB);
-
-		// Check if the key exists in the response
-		if (!idvidResponse.containsKey(dateOfBirthFieldName)) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::getDateOfBirthFromIdrepo():: fieldName does not exist in idvidResponse : {} ", dateOfBirthFieldName);
-			return null;
-		}
-
-		// Safely get the value
-		String dobValue = JsonUtil.getJSONValue(idvidResponse, dateOfBirthFieldName);
-
-		// Check if the value itself is null
-		if (dobValue == null) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::getDateOfBirthFromIdrepo():: fieldValue does not exist in idvidResponse : {} ", dateOfBirthFieldName);
-			return null;
-		}
-
-		LocalDate dob = parseToLocalDate(dobValue, dobFormat);
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getDateOfBirthFromIdrepo()::exit with dob");
-		return dob;
-	}
-
-	/**
-	 * Calculates the age of an applicant at the time of the last packet processing
-	 *
-	 * @param dateOfBirth
-	 * @param lastPacketProcessingDate
-	 * @param rid
-	 * @return
-	 */
-	public int calculateAgeAtLastPacketProcessing(LocalDate dateOfBirth, LocalDate lastPacketProcessingDate, String rid) {
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::calculateAgeAtLastPacketProcessing():: entry");
-
-		// Calculate the period between the two dates
-		Period period = Period.between(dateOfBirth, lastPacketProcessingDate);
-
-		// Extract years from the period
-		int ageInYears = period.getYears();
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::calculateAgeAtLastPacketProcessing():: exit");
-
-		// Return age in years (as per the original method signature)
-		return ageInYears;
-	}
-
-	/**
-	 * Obtain the IdVidMetadata for applicant
-	 *
-	 * @param individualId
-	 * @param idType
-	 * @return
-	 * @throws IOException
-	 * @throws ApisResourceAccessException
-	 */
-	public IdVidMetadataResponse getIdVidMetadata(String individualId, String idType) throws IOException, ApisResourceAccessException {
-		IdVidMetadataRequest idVidMetadataRequest = new IdVidMetadataRequest();
-		idVidMetadataRequest.setIndividualId(individualId);
-		idVidMetadataRequest.setIdType(idType);
-		IdVidMetadataResponse idVidMetadataResponse = idRepoService.searchIdVidMetadata(idVidMetadataRequest);
-		return idVidMetadataResponse;
-	}
-
-	/**
-	 * Retrieves the packet creation date for the given RID from the sync registration table
-	 *
-	 * @param rid
-	 * @return
-	 */
-	public LocalDate getPacketCreatedDateFromSyncRegistration(String rid) {
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-				LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getPacketCreatedDateFromSyncRegistration():: entry");
-
-		List<SyncRegistrationEntity> registrations = syncRegistrationRepository.findByRegistrationId(rid);
-		if (registrations == null || registrations.isEmpty()) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-			LoggerFileConstant.REGISTRATIONID.toString(), rid,
-			"utility::getPacketCreatedDateFromSyncRegistration():: no SyncRegistration records found");
-			return null;
-		}
-
-		// Fetch latest packetId based on createOn
-		String packetId = registrations.stream()
-				.max(Comparator.comparing(SyncRegistrationEntity::getCreateDateTime)) // latest record
-				.map(SyncRegistrationEntity::getPacketId) // extract packetId
-				.orElse(null); // if no records, return null
-
-		LocalDate packetCreatedDate = extractPacketCreatedDate(packetId);
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-				LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getPacketCreatedDateFromSyncRegistration():: exit with packetCreatedDate: " + packetCreatedDate);
-
-		return packetCreatedDate;
-	}
-
-	private LocalDate extractPacketCreatedDate(String id) {
-		String packetCreatedDateFormat = "yyyyMMddHHmmss";
-
-		if (id == null || id.length() < 14) {
-			regProcLogger.debug("Provided id is null or shorter than 14 characters. Cannot extract packet creation date. id : {} " , id);
-			return null;
-		}
-
-		String dateStr = id.substring(id.length() - 14);
-		return parseToLocalDate(dateStr, packetCreatedDateFormat);
-	}
-
-
-	/**
-	 * Wrapper that returns LocalDate only (ignores time)
-	 *
-	 * @param dateString
-	 * @param dateFormat
-	 * @return
-	 */
-	public LocalDate parseToLocalDate(String dateString, String dateFormat) {
-
-		// Log the incoming parameters for debugging
-		regProcLogger.debug("Attempting to parse date: {} with format: {}", dateString, dateFormat);
-
-		LocalDateTime ldt = null;
-
-		try {
-			// Parse the date string into LocalDateTime
-			ldt = parseUTCToLocalDateTime(dateString, dateFormat);
-
-			// Perform validation if parsing was successful
-			if (ldt != null) {
-				LocalDateTime now = LocalDateTime.now();
-
-				// Check if date is in the future
-				if (after(ldt, now)) {
-					regProcLogger.error("Parsed LocalDateTime occurs in the future: {}", ldt);
-					return null;
-				}
-
-				// Check if date is older than 100 years
-				LocalDateTime hundredYearsAgo = now.minusYears(200);
-				if (before(ldt, hundredYearsAgo)) {
-					regProcLogger.error("Date is older than 200 years : {}", ldt);
-					return null;
-				}
-			}
-		} catch (io.mosip.kernel.core.exception.ParseException e) {
-				regProcLogger.debug("Failed to parse date: {} with format: {}", dateString, dateFormat, e);
-				return null;
-		}
-
-		return (ldt != null) ? ldt.toLocalDate() : null;
-	}
-
-	/**
-	 * Extracts the packet creation date from the given Registration ID (RID) by interpreting its last 14 digits as a timestamp in the format {@code yyyyMMddHHmmss}
-	 *
-	 * @param rid
-	 * @return
-	 */
-	public LocalDate getPacketCreatedDateFromRid(String rid) {
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-				LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getPacketCreatedDateFromRid():: entry");
-
-		LocalDate packetCreatedDate = extractPacketCreatedDate(rid);
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(),
-				LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getPacketCreatedDateFromRid():: exit with packetCreatedDate: " + packetCreatedDate);
-
-		return packetCreatedDate;
-	}
-
-	/**
-	 * Computes the approximate packet creation date (for the packet that updated the identity)
-	 *
-	 * @param idVidMetadataDTO
-	 * @param processingRid
-	 * @return
-	 */
-	public LocalDate computePacketCreatedFromIdentityUpdate(IdVidMetadataResponse idVidMetadataDTO, String processingRid) {
-
-		String packetCreatedDateTimeIsoFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
-		String updatedOn = idVidMetadataDTO.getUpdatedOn();
-		regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), processingRid,
-				"computePacketCreatedFromIdentityUpdate() :: expectedPacketProcessingDurationHours: {}", expectedPacketProcessingDurationHours);
-
-		if (updatedOn == null) {
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), processingRid,
-					"computePacketCreatedFromIdentityUpdate() :: updatedOn is null for RID hence falling back to createdOn.");
-			updatedOn = idVidMetadataDTO.getCreatedOn();
-		}
-
-		try {
-			LocalDateTime referenceDateTime = parseUTCToLocalDateTime(updatedOn, packetCreatedDateTimeIsoFormat);
-			LocalDateTime adjustedDateTime = referenceDateTime.minusHours(expectedPacketProcessingDurationHours);
-
-			return adjustedDateTime.toLocalDate();
-
-		} catch (io.mosip.kernel.core.exception.ParseException e) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), processingRid,
-					"computePacketCreatedFromIdentityUpdate() :: Failed to parse date {} using format {}. Exception: {}",
-					updatedOn, packetCreatedDateTimeIsoFormat, e.getMessage());
-
-		} catch (Exception e) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), processingRid,
-					"computePacketCreatedFromIdentityUpdate() :: Unexpected error while computing packet creation date. Exception: {}",
-					e.getMessage());
-		}
-		return null;
-	}
-
-	/**
-	 * Get the BiometricRecord from the IdRepo for a given UIN
-	 *
-	 * @param uin
-	 * @param rid
-	 * @return
-	 * @throws ApisResourceAccessException
-	 * @throws IOException
-	 */
-	public BiometricRecord getBiometricRecordfromIdrepo(String uin, String rid) throws ApisResourceAccessException, IOException {
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::getBiometricRecordfromIdrepo():: entry");
-
-		// Step 1: Retrieve all documents from IDRepo
-		List<Documents> documents = retrieveIdrepoDocument(uin);
-
-		// Step 2: Load mapping JSON and extract the label for individual biometrics
-		JSONObject regProcessorIdentityJson = getRegistrationProcessorMappingJson(MappingJsonConstants.IDENTITY);
-		String individualBiometricsLabel = JsonUtil.getJSONValue(
-				JsonUtil.getJSONObject(regProcessorIdentityJson, MappingJsonConstants.INDIVIDUAL_BIOMETRICS),
-				MappingJsonConstants.VALUE);
-
-		// Step 3: Find the biometric document
-		String biometricDoc = null;
-		for (Documents doc : documents) {
-			if (doc.getCategory() != null && doc.getCategory().equalsIgnoreCase(individualBiometricsLabel)) {
-				biometricDoc = doc.getValue();
-				break;
-			}
-		}
-
-		if (biometricDoc == null) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"No biometric document found in IDRepo");
-			throw new BiometricClassificationException(PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getMessage());
-		}
-
-		try {
-			// Step 4: Decode and convert to BiometricRecord
-			byte[] bi = CryptoUtil.decodeURLSafeBase64(biometricDoc);
-
-			BIR birs = CbeffValidator.getBIRFromXML(bi);
-
-			BiometricRecord biometricRecord = new BiometricRecord();
-			// Copy "others" metadata if present
-			if(birs.getOthers() != null) {
-				HashMap<String, String> others = new HashMap<>();
-				birs.getOthers().entrySet().forEach(e -> {
-					others.put(e.getKey(), e.getValue());
-				});
-				biometricRecord.setOthers(others);
-			}
-
-			biometricRecord.setSegments(birs.getBirs());
-
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::getBiometricRecordfromIdrepo():: exit");
-			return biometricRecord;
-
-		} catch (Exception e) {
-			// Any other error during XML parsing or processing
-			String errorMsg = "Unexpected error while getting biometric document";
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
-					rid, errorMsg, e);
-			throw new BiometricClassificationException(
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getMessage());
-		}
-	}
-
-	/**
-	 * Checks whether all biometric segments are marked as exceptions excluding FACE and EXCEPTION_PHOTO types, which cannot be marked as exceptions
-	 *
-	 * @param birs
-	 * @param rid
-	 * @return
-	 * @throws BiometricClassificationException
-	 */
-	public boolean allBiometricHaveException(List<BIR> birs, String rid) throws BiometricClassificationException {
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::allBiometricHaveException():: entry");
-
-		if (birs == null || birs.isEmpty()) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::allBiometricHaveException():: Biometric list is null or empty");
-			throw new BiometricClassificationException(
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getMessage());
-		}
-
-		boolean hasOthers = hasBiometricWithOthers(birs);
-		regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::allBiometricHaveException():: hasOthers: {} ",hasOthers);
-
-		for (BIR bir : birs) {
-			BiometricType type = bir.getBdbInfo().getType().get(0);
-			boolean isFaceOrExceptionPhoto = type == BiometricType.FACE || type == BiometricType.EXCEPTION_PHOTO;
-
-			regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::allBiometricHaveException():: Checking biometric type: {} ",type);
-
-			if (hasOthers) {
-				if (!isFaceOrExceptionPhoto) {
-					String exceptionValue = bir.getOthers().get(EXCEPTION);
-					regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-							"utility::allBiometricHaveException()::  Biometric type: {},  exceptionValue: {}", type, exceptionValue);
-
-					if (exceptionValue == null || !exceptionValue.equalsIgnoreCase(TRUE)) {
-						regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-								"utility::allBiometricHaveException():: Biometric type: {} does not have exception", type);
-						return false;
-					}
-				}
-			} else {
-				if (!isFaceOrExceptionPhoto) {
-					regProcLogger.info(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-							"utility::allBiometricHaveException():: Biometric type: {} does not have exception", type);
-					return false;
-				}
-			}
-		}
-
-		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-				"utility::allBiometricHaveException():: exit");
-		return true;
-	}
-
-	/**
-	 * Checks whether any biometric record in the given list contains non-empty "others" data,
-	 * 	which can be used to determine if the biometric was generated using the new or old version
-	 *
-	 * @param bir
-	 * @return
-	 */
-	public boolean hasBiometricWithOthers(List<BIR> bir){
-		if (bir == null || bir.isEmpty()) {
-			return false;
-		}
-		return bir.stream()
-				.anyMatch(bi -> bi.getOthers() != null && !bi.getOthers().isEmpty());
-	}
-
-	/**
-	 * Checks whether all biometric segments are marked as exceptions for a registration Id
-	 *
-	 * @param rid
-	 * @param registrationType
-	 * @param stageName
-	 * @return
-	 * @throws BiometricClassificationException
-	 */
-	public boolean allBiometricHaveException(String rid, String registrationType, ProviderStageName stageName) throws BiometricClassificationException {
-		try {
-			String uin = utility.getUIn(rid, registrationType, stageName);
-			BiometricRecord bm = getBiometricRecordfromIdrepo(uin, rid);
-			return allBiometricHaveException(bm.getSegments(), rid);
-		} catch (Exception e) {
-			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(), rid,
-					"utility::allBiometricHaveException():: Error while classifying biometric exceptions", e);
-			throw new BiometricClassificationException(
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getCode(),
-					PlatformErrorMessages.RPR_BDD_UNABLE_TO_FETCH_BIOMETRIC_INFO.getMessage());
-		}
-	}
-
-	/**
-	 * Retrieves the packet creation date from the packet for a given registrationId
-	 *
-	 * @param rid
-	 * @param process
-	 * @param stageName
-	 * @return
-	 * @throws PacketManagerException
-	 * @throws ApisResourceAccessException
-	 * @throws IOException
-	 * @throws JsonProcessingException
-	 */
-	public String retrieveCreatedDateFromPacket(String rid, String process, ProviderStageName stageName)
-			throws PacketManagerException, ApisResourceAccessException, IOException, JsonProcessingException {
-
-		Map<String, String> metaInfo = packetManagerService.getMetaInfo(rid, process, stageName);
-		String packetCreatedDateTime = metaInfo.get(JsonConstant.CREATIONDATE);
-
-		if (packetCreatedDateTime != null && !packetCreatedDateTime.isEmpty()) {
-			return packetCreatedDateTime;
-		}
-
-		regProcLogger.error(
-				LoggerFileConstant.SESSIONID.toString(),
-				LoggerFileConstant.REGISTRATIONID.toString(),
-				" -- " + rid,
-				PlatformErrorMessages.RPR_PVM_PACKET_CREATED_DATE_TIME_EMPTY_OR_NULL.getMessage()
-		);
-
-		return null;
 	}
 }
